+++ Plugin.cs
@@ -20,7 +20,7 @@ public class SecondaryAttacksPlugin : BaseUnityPlugin
     internal const string MagicPluginGuid = "blacks7ar.MagicPlugin";
     internal const string QuickstepPluginGuid = "shudnal.Quickstep";
     internal const string ModName = "SecondaryAttacks";
-    internal const string ModVersion = "1.1.4";
+    internal const string ModVersion = "1.1.5";
     internal const string Author = "sighsorry";
     private const string ModGUID = $"{Author}.{ModName}";
     private static string ConfigFileName = $"{ModGUID}.cfg";